reorder checks

+++ pkg/util/kernel/headers/find_headers.go
@@ -393,51 +393,50 @@ func getDownloadedHeaderDirs(headerDownloadDir string) []string {
 
 var errInvalidTempDirectory = errors.New("invalid system-probe temp directory")
 
+func checkTempDirPermissions(dir string) error {
+	sfi, statErr := os.Lstat(dir)
+	if statErr != nil {
+		return statErr
+	}
+
+	if !sfi.IsDir() {
+		return fmt.Errorf("%w: %s is not a directory", errInvalidTempDirectory, dir)
+	}
+	if sfi.Mode()&fs.ModeSymlink != 0 {
+		return fmt.Errorf("%w: %s is not allowed to be a symlink", errInvalidTempDirectory, dir)
+	}
+
+	stat, ok := sfi.Sys().(*syscall.Stat_t)
+	if !ok {
+		return fmt.Errorf("unable to verify permissions of %s", dir)
+	}
+	if stat.Uid != 0 || stat.Gid != 0 || sfi.Mode().Perm()&os.FileMode(0022) != 0 {
+		return fmt.Errorf("%w: %s has incorrect permissions: user=%d, group=%d, permissions=%s", errInvalidTempDirectory, dir, stat.Uid, stat.Gid, sfi.Mode().Perm())
+	}
+	return nil
+}
+
 func getSysfsHeaderDirs(tmpDir string, v kernel.Version) ([]string, error) {
 	sysprobeTmpPath := filepath.Join(tmpDir, "system-probe")
 	tmpPath := filepath.Join(sysprobeTmpPath, fmt.Sprintf("linux-headers-%s", v))
 
-	sfi, statErr := os.Lstat(sysprobeTmpPath)
-	if statErr != nil {
-		if errors.Is(statErr, fs.ErrNotExist) {
-			if err := os.MkdirAll(sysprobeTmpPath, 0755); err != nil {
-				return nil, fmt.Errorf("create system-probe temp directory: %s", err)
-			}
-		} else {
-			return nil, fmt.Errorf("stat %s: %s", sysprobeTmpPath, statErr)
+	if fi, err := os.Stat(tmpPath); err == nil && fi.IsDir() {
+		if err := checkTempDirPermissions(sysprobeTmpPath); err != nil {
+			return nil, err
 		}
-	} else {
-		if !sfi.IsDir() {
-			return nil, fmt.Errorf("%w: %s is not a directory", errInvalidTempDirectory, sysprobeTmpPath)
+		hv, err := getHeaderVersion(tmpPath)
+		if err != nil {
+			// remove tmp dir if it errors
+			_ = os.RemoveAll(tmpPath)
+			return nil, fmt.Errorf("unable to verify headers version: %w", err)
 		}
-		if sfi.Mode()&fs.ModeSymlink != 0 {
-			return nil, fmt.Errorf("%w: %s is not allowed to be a symlink", errInvalidTempDirectory, sysprobeTmpPath)
+		if hv != v {
+			// remove tmp dir if it fails to validate
+			_ = os.RemoveAll(tmpPath)
+			return nil, fmt.Errorf("header version %s does not match expected host version %s", v, hv)
 		}
-
-		stat, ok := sfi.Sys().(*syscall.Stat_t)
-		if !ok {
-			return nil, fmt.Errorf("unable to verify permissions of %s", sysprobeTmpPath)
-		}
-		if stat.Uid != 0 || stat.Gid != 0 || sfi.Mode().Perm()&os.FileMode(0022) != 0 {
-			return nil, fmt.Errorf("%w: %s has incorrect permissions: user=%d, group=%d, permissions=%s", errInvalidTempDirectory, sysprobeTmpPath, stat.Uid, stat.Gid, sfi.Mode().Perm())
-		}
-
-		fi, err := os.Stat(tmpPath)
-		if err == nil && fi.IsDir() {
-			hv, err := getHeaderVersion(tmpPath)
-			if err != nil {
-				// remove tmp dir if it errors
-				_ = os.RemoveAll(tmpPath)
-				return nil, fmt.Errorf("unable to verify headers version: %w", err)
-			}
-			if hv != v {
-				// remove tmp dir if it fails to validate
-				_ = os.RemoveAll(tmpPath)
-				return nil, fmt.Errorf("header version %s does not match expected host version %s", v, hv)
-			}
-			log.Debugf("found valid kernel headers at %s", tmpPath)
-			return []string{tmpPath}, nil
-		}
+		log.Debugf("found valid kernel headers at %s", tmpPath)
+		return []string{tmpPath}, nil
 	}
 
 	if !sysfsHeadersExist() {
@@ -450,6 +449,15 @@ func getSysfsHeaderDirs(tmpDir string, v kernel.Version) ([]string, error) {
 		}
 	}
 
+	if _, err := os.Lstat(sysprobeTmpPath); errors.Is(err, fs.ErrNotExist) {
+		if err := os.MkdirAll(sysprobeTmpPath, 0755); err != nil {
+			return nil, fmt.Errorf("create system-probe temp directory: %s", err)
+		}
+	}
+
+	if err := checkTempDirPermissions(sysprobeTmpPath); err != nil {
+		return nil, err
+	}
 	if err := archive.TarXZExtractAll(sysfsHeadersPath, tmpPath); err != nil {
 		return nil, fmt.Errorf("unable to extract kernel headers: %w", err)
 	}
